+++ addon/locales/ru/translations.js
@@ -131,6 +131,7 @@ export default {
       'filter-by-any-match-placeholder': 'Поиск...',
       'create-setting-title': 'Новая настройка',
       'use-setting-title': 'Применить',
+      'export-title': 'Выгрузить',
       'edit-setting-title': 'Редактировать',
       'remove-setting-title': 'Удалить',
       'set-default-setting-title': 'Установка по умолчанию',
@@ -145,6 +146,7 @@ export default {
 
     'colsconfig-dialog-content': {
       'title': 'Настроить отображение столбцов',
+      'export-title': 'Экспорт в Excel',
       'dont-show-columns': 'Не отображать столбцы',
       'columns-order': 'Определить порядок столбцов',
       'column-name': 'Название столбца',
@@ -155,6 +157,7 @@ export default {
       'setting-name': 'Название настройки',
       'enter-setting-name': 'Введите название настройки',
       'use': 'Применить',
+      'export': 'Выгрузить',
       'use-without-save': 'Применить данные установки без сохранения в настройке ',
       'save': 'Сохранить',
       'have-errors': 'При сохранении настройки возникли ошибки: ',
@@ -162,6 +165,8 @@ export default {
       'is-saved': ' сохранена',
       'col-width-on': 'Включить настройку ширины столбцов',
       'per-page': 'Записей на странице',
+      'det-separate-rows': 'Поля списков в отдельные строки',
+      'det-separate-cols': 'Поля списков в отдельные столбцы',
       'unresizable': 'Фиксированная ширина'
     },